feat: refactor server tool registration and add utility functions for fetching data
