+++ .eslintrc
@@ -1,5 +1,9 @@
 {
-    "extends": ["plugin:prettier/recommended"],
+    "parser": "@typescript-eslint/parser",
+    "extends": [
+        "plugin:@typescript-eslint/recommended",
+        "plugin:prettier/recommended"
+    ],
     "parserOptions": {
         "ecmaVersion": "latest",
         "sourceType": "module"
