+++ src/mcp-server/prompts/index.ts
@@ -1,0 +1,6 @@
+import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
+import { registerAnalysisPrompts } from './analysis.js';
+
+export function registerPrompts(server: McpServer) {
+    registerAnalysisPrompts(server);
+}
