+++ src/mcp-server/tools/index.ts
@@ -1,0 +1,10 @@
+import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
+import { registerLogTools } from './logs.js';
+import { registerQueryTools } from './queries.js';
+import { registerGeneralTools } from './general.js';
+
+export function registerTools(server: McpServer) {
+    registerLogTools(server);
+    registerQueryTools(server);
+    registerGeneralTools(server);
+}
