+++ src/mcp-server/tools/queries.ts
@@ -1,0 +1,26 @@
+import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
+import { z } from 'zod';
+import { fetchData } from '../utils.js';
+
+export function registerQueryTools(server: McpServer) {
+    server.registerTool(
+        'get_queries',
+        {
+            description: 'Get captured SQL queries',
+            inputSchema: {
+                limit: z.number().optional().describe('Limit the number of queries returned (default: 50)')
+            }
+        },
+        async ({ limit = 50 }) => {
+            const data = await fetchData('queries');
+            if (data.error) return { content: [{ type: 'text', text: `Error: ${data.error}` }] };
+
+            let queries = Array.isArray(data) ? data : Object.values(data);
+            queries = queries.slice(-limit);
+
+            return {
+                content: [{ type: 'text', text: JSON.stringify(queries, null, 2) }]
+            };
+        }
+    );
+}
